Add custom appender to the nav context
