+++ packages/edit-navigation/src/filters/index.js
@@ -1,12 +1,10 @@
 /**
  * Internal dependencies
  */
-import addNavigationEditorCustomAppender from './add-navigation-editor-custom-appender';
 import addMenuNameEditor from './add-menu-name-editor';
 import removeSettingsUnsupportedFeatures from './remove-settings-unsupported-features';
 
 export const addFilters = () => {
-	addNavigationEditorCustomAppender();
 	addMenuNameEditor();
 	removeSettingsUnsupportedFeatures();
 };
